move logic to base.py
